requirements.txt
boto3==1.20.26
botocore==1.23.26
Pillow==8.4.0
pyactiveresource==2.2.2
python-dateutil==2.8.2
s3transfer==0.5.0
ShopifyAPI==9.0.0
urllib3==1.26.7
